support local dev, move stack tests
