+++ experimental/rust-rocket/image/Dockerfile-stack
@@ -1,32 +1,35 @@
 FROM appsody/rust:0.1
 
 RUN rustup default nightly \
- && rustup update
+ && rustup update \
+ && cargo install cargo-edit
 
 ENV CARGO_HOME=/usr/local/cargo/deps
+ENV CARGO_TARGET_DIR=$CARGO_HOME/target
 
-ENV APPSODY_MOUNTS=.:/project/user-app
+ENV APPSODY_MOUNTS=/Cargo.toml:/Cargo.toml;/src:/project/user-app/src/
 ENV APPSODY_DEPS=$CARGO_HOME
 
 ENV APPSODY_WATCH_DIR="/project/user-app"
+ENV APPSODY_WATCH_IGNORE_DIR="/project/user-app/.appsody"
 ENV APPSODY_WATCH_REGEX="(^.*.rs$)"
 
 ENV ROCKET_ADDRESS=0.0.0.0 
 ENV ROCKET_PORT=8000 
 
+ENV APPSODY_PREP="cp /Cargo.toml /project/user-app/Cargo.toml && cargo add -q appsody-rocket --path=../deps"
+
 ENV APPSODY_RUN="cargo run"
 ENV APPSODY_RUN_ON_CHANGE="cargo run"
 ENV APPSODY_RUN_KILL=true
 
-ENV APPSODY_TEST="cargo test"
+ENV APPSODY_TEST="cargo test -p appsody-rocket && cargo test"
 
 COPY ./LICENSE /licenses
 COPY ./project /project
 COPY ./config /config
 
 WORKDIR /project/user-app
-
-RUN cargo fetch
 
 ENV PORT=8000
