+++ experimental/rust-rocket/image/project/deps/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "rust-rocket"
+name = "appsody-rocket"
 version = "0.1.0"
-authors = ["Name <name@email.com>"]
+description = "Package containing dependencies for the Appsody Rocket stack"
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
