+++ experimental/rust-rocket/templates/simple/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-rust-rocket = { path = "../../project" }
+appsody-rocket = { path = ".appsody" }
 rocket = "0.4.2"
 
 [[bin]]
