+++ experimental/rust-rocket/templates/simple/src/tests.rs
@@ -9,17 +9,3 @@ fn hello_world() {
     assert_eq!(response.status(), Status::Ok);
     assert_eq!(response.body_string(), Some("Hello from Appsody!".into()));
 }
-
-#[test]
-fn metrics() {
-    let client = Client::new(rocket()).unwrap();
-    let response = client.get("/metrics").dispatch();
-    assert_eq!(response.status(), Status::Ok)
-}
-
-#[test]
-fn random() {
-    let client = Client::new(rocket()).unwrap();
-    let response = client.get("/random").dispatch();
-    assert_eq!(response.status(), Status::NotFound);
-}
